pyproject.toml
[project]
name = "syntaks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-contentsafety>=1.0.0",
    "azure-identity>=1.21.0",
    "azure-keyvault>=4.2.0",
    "bertopic>=0.17.0",
    "datamapplot>=0.3.0",
    "gradio>=5.29.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "matplotlib>=3.10.1",
    "nbformat>=5.10.4",
    "nbstripout>=0.8.1",
    "openai>=1.76.2",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "pydantic>=2.11.4",
    "pydantic-ai>=0.1.9",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "transformers>=4.51.3",
    "umap-learn>=0.5.7",
    "wordcloud>=1.9.4",
    "xgboost>=3.0.0",
]
